python-ai-engine/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
langgraph
langchain
langchain-openai
langchain-groq
pydantic
python-dotenv
requests
